+++ breeze-dsl/src/main/kotlin/com/windea/breezeframework/dsl/mermaid/MermaidDsl.kt
@@ -19,16 +19,17 @@ private annotation class MermaidDsl
 abstract class Mermaid : DslBuilder
 
 /**Mermaid Dsl的配置。*/
+@ReferenceApi("[Mermaid](https://mermaidjs.github.io)")
 @MermaidDsl
 object MermaidConfig : DslConfig {
 	/**缩进长度。*/
 	var indentSize = 4
 		set(value) = run { field = value.coerceIn(-2, 8) }
 	/**是否使用双引号。*/
-	var useDoubleQuote: Boolean = true
+	var preferDoubleQuote: Boolean = true
 	
 	@PublishedApi internal val indent get() = if(indentSize <= -1) "\t" * indentSize else " " * indentSize
-	@PublishedApi internal val quote get() = if(useDoubleQuote) "\"" else "'"
+	@PublishedApi internal val quote get() = if(preferDoubleQuote) '"' else '\''
 }
 
 /**Mermaid Dsl的元素。*/