+++ breeze-generator/src/main/kotlin/com/windea/breezeframework/generator/code/SqlGenerator.kt
@@ -43,7 +43,7 @@ object SqlGenerator : Generator {
 			insert into $tableName ($columnNamesSnippet) values
 			${table.joinToString(",\n", "", ";\n") { data ->
 				val columnsSnippet = data.values.joinToString {
-					it.toString().wrapQuote("'").escape()
+					it.toString().wrapQuote('\'').escape()
 				}
 				
 				"""  ($columnsSnippet)"""